third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/nonces/script-enforce-blocked-expected.txt

CONSOLE ERROR: line 37: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-rwG0Y6oerUTSWQ5J1R+4xK3lr2biWxqq601ynqg0uJQ='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 40: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-rwG0Y6oerUTSWQ5J1R+4xK3lr2biWxqq601ynqg0uJQ='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 43: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-rwG0Y6oerUTSWQ5J1R+4xK3lr2biWxqq601ynqg0uJQ='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 46: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-rwG0Y6oerUTSWQ5J1R+4xK3lr2biWxqq601ynqg0uJQ='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 49: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'". Either the 'unsafe-inline' keyword, a hash ('sha256-rwG0Y6oerUTSWQ5J1R+4xK3lr2biWxqq601ynqg0uJQ='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: Refused to load the script 'https://evil.example.test/yay1.js' because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'".

CONSOLE ERROR: Refused to load the script 'https://evil.example.test/yay2.js' because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'".

CONSOLE ERROR: Refused to load the script 'https://evil.example.test/yay3.js' because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'".

CONSOLE ERROR: Refused to load the script 'https://evil.example.test/yay4.js' because it violates the following Content Security Policy directive: "script-src 'self' 'nonce-abc'".

CONSOLE MESSAGE: line 30: 1 : 37 : inline
CONSOLE MESSAGE: line 30: 2 : 40 : inline
CONSOLE MESSAGE: line 30: 3 : 43 : inline
CONSOLE MESSAGE: line 30: 4 : 46 : inline
CONSOLE MESSAGE: line 30: 5 : 49 : inline
CONSOLE MESSAGE: line 30: 6 : 0 : https://evil.example.test/yay1.js
CONSOLE MESSAGE: line 30: 7 : 0 : https://evil.example.test/yay2.js
CONSOLE MESSAGE: line 30: 8 : 0 : https://evil.example.test/yay3.js
CONSOLE MESSAGE: line 30: 9 : 0 : https://evil.example.test/yay4.js
This is a testharness.js-based test.
PASS Unnonced scripts generate reports. 
PASS Blocks without correct nonce do not execute, and generate violation reports 
Harness: the test ran to completion.
